Extract w&b code into module
